start of icon support for data model/cm
git-svn-id: https://subversion.assembla.com/svn/smart-cs/branches/emily_iconsupport@7040 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.report/src/org/wcs/smart/report/internal/ui/viewer/ReportView.java
@@ -174,7 +174,7 @@ public class ReportView implements IReportListener{
 			ReportPlugIn.log(e.getMessage(), e);
 		}
 		
-		if (tempReportDocument != null) {
+		if (tempReportDocument != null && Files.exists(tempReportDocument)) {
 			try {
 				Files.delete(tempReportDocument);
 			}catch (Exception ex) {
